Propose cachability is a key success criteria
Sorry to be jumping in, but I believe that cachability is absolutely essential when it comes to acl treatment. We have to check access for every resource, and so, the acl check adds latency before first byte. Failure in this area is a potential show-stopper, so I would like to have it very high on the agenda.