Update 2021-04-28.md
I have refactored the discussion a bit more into a dialogue, filled out comments I made, and linked to examples I posted over the past
few months which I alluded to in the conversation, but did not have time to fill in. 
Does that still work for you @matthieubosquet ?

The last part still needs filling in by Ericp.